renaming, docstrings
