+++ asab/web/auth/__init__.py
@@ -1,6 +1,6 @@
 from .decorator import require, noauth
 from .service import AuthService
-from .authz import Authorization
+from .authorization import Authorization
 
 __all__ = (
 	"AuthService",
